pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blabberer"
version = "0.1.0"
description = "Generate lip sync mouth shape sprites from any face image using AI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "HistOracle", email = "team@historacle.ai" },
]
keywords = ["lip-sync", "viseme", "mouth-shapes", "sprites", "animation", "ai", "blabberer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "Pillow>=10.0",
    "rembg>=2.0",
    "mediapipe>=0.10",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
cpu = ["onnxruntime>=1.14"]
gpu = ["onnxruntime-gpu>=1.14"]

[project.scripts]
blabberer = "blabberer.cli:main"

[project.urls]
Homepage = "https://github.com/historacle-ai/blabberer"
Repository = "https://github.com/historacle-ai/blabberer"
Issues = "https://github.com/historacle-ai/blabberer/issues"

[tool.hatch.build.targets.wheel]
packages = ["blabberer"]

[tool.hatch.build.targets.sdist]
include = ["blabberer/", "reference/", "README.md", "LICENSE"]
